Add files via upload
This version uses a red colored object as a replacement for the hand because skin-based detection using only classical computer vision often detects other body parts such as the face and arm due to similar color and contour properties. This causes unstable and inaccurate tracking.

To avoid this issue and improve reliability, a bright red object is used as the tracked marker, resulting in much more stable and accurate real-time detection.